requirements.txt
langchain>=0.3
langchain-core
langchain-google-genai
langchain-community
langchain-chroma
langchain-text-splitters
langgraph
chromadb
pdfplumber
pymupdf
python-docx
streamlit
duckduckgo-search
python-dotenv
pillow
typing-extensions
google-generativeai
